ChessGame.cpp
The Game have been designed in C++ keeping in mind the object oriented practices of the  C++ . The Base class i.e Gamepiece Class has 2 pure virtual functions i.e these functions has to be over-ridden who even extends the base class or else the extending class also becomes the abstract class. Getpiece & AreSquareLegal are 2 virtual functions in the base class  . Gamepiece is an abstract class as it has pure virtual functions .

The Gamepiece class as two pure virtual functions: GetPiece() and AreSquaresLegal(). The function GetPiece() is overridden by each piece class to return the char that signifies each piece type: 'P' = pawn, 'N' = knight, 'R' = rook, 'B' = bishop, 'Q' = queen, and 'K' = king. The function AreSquaresLegal() is overridden to test the source and destination locations versus the movement rules for the particular piece.

The board is numbered on the left and bottom from 1 to 8 (as shown above) so that locations ca be specified as a two-digit number and moves can specified as a pair of two-digit numbers. The green squares show the legal moves for a knight. So, that 54 to 42 is a legal move and will return true if passed into the knight's AreSquaresLegal() function.

he main game board is an 8 by 8 array of pointers to piece objects. This is where the polymorphism comes in. Any square can hold any piece or none at all. If we want to know whether a piece can make a certain move, we just pass in the move to the IsLegalMove() function in the piece class. Then correct function is called for any of the pieces: pawn, knight, bishop, rook, queen, and king. However, all we know when the function is called is that it is a piece.

Now we will go through the each of the functions so , that implementations of the functions become clear .

==>  Keys to sysmbols used 

 * = white space

 Blank space = black space

 WP = White pawn &  BP = Black pawn

 WN = White Knight & BN = Black Knight

 WB = White Bishop & BB = Black Bishop

 WR = White Rook & BR = Black Rook

 WQ = White Queen & BQ = Black Queen

 WK = White King & BK =Black King

Rule for move is :

Move by selecting row & column to another valid location using row & column

Valid location is checked by the code , invalid location is entered then user is prompted to enter once again